+++ src/preprocessing/passes/theory_rewrite_eq.h
