+++ src/preprocessing/preprocessing_pass_registry.cpp
@@ -25,7 +25,7 @@
 #include "base/output.h"
 #include "preprocessing/passes/ackermann.h"
 #include "preprocessing/passes/apply_substs.h"
-#include "preprocessing/passes/arith_rewrite_eq.h"
+#include "preprocessing/passes/theory_rewrite_eq.h"
 #include "preprocessing/passes/bool_to_bv.h"
 #include "preprocessing/passes/bv_abstraction.h"
 #include "preprocessing/passes/bv_eager_atoms.h"
@@ -150,7 +150,7 @@ PreprocessingPassRegistry::PreprocessingPassRegistry()
   registerPassInfo("bool-to-bv", callCtor<BoolToBV>);
   registerPassInfo("ho-elim", callCtor<HoElim>);
   registerPassInfo("fun-def-fmf", callCtor<FunDefFmf>);
-  registerPassInfo("arith-rewrite-eq", callCtor<ArithRewriteEq>);
+  registerPassInfo("theory-rewrite-eq", callCtor<TheoryRewriteEq>);
 }
 
 }  // namespace preprocessing
